Move gas model to shared and rename http solver module
