+++ crates/shared/src/http_solver.rs
@@ -4,6 +4,7 @@ use reqwest::{Client, Url};
 use std::time::Duration;
 
 pub mod model;
+pub mod gas_model;
 
 /// Implements an abstract HTTP solver API, can be mocked, instrumented, etc.
 #[mockall::automock]
