+++ crates/shared/src/http_solver/model.rs
