+++ crates/solver/src/solver.rs
@@ -16,7 +16,7 @@ use oneinch_solver::OneInchSolver;
 use paraswap_solver::ParaswapSolver;
 use reqwest::{Client, Url};
 use shared::balancer_sor_api::DefaultBalancerSorApi;
-use shared::http_solver_api::{DefaultHttpSolverApi, SolverConfig};
+use shared::http_solver::{DefaultHttpSolverApi, SolverConfig};
 use shared::zeroex_api::ZeroExApi;
 use shared::{
     baseline_solver::BaseTokens, conversions::U256Ext, token_info::TokenInfoFetching, Web3,
