+++ crates/solver/src/solver/http_solver/settlement.rs
@@ -8,7 +8,7 @@ use anyhow::{anyhow, Result};
 use ethcontract::Bytes;
 use model::order::OrderKind;
 use primitive_types::{H160, U256};
-use shared::http_solver_api::model::*;
+use shared::http_solver::model::*;
 use std::collections::{hash_map::Entry, HashMap};
 
 // To send an instance to the solver we need to identify tokens and orders through strings. This
